Added logic for phaser capacitors

+++ src/main/java/com/sfb/systems/PerformanceData.java
@@ -7,19 +7,19 @@ public class PerformanceData {
 	// Performance statistics for this spaceframe.
 	
 	// Base values
-	private double movementCost	= 0;
-	private boolean nimble		= false;	// Nimble ships can do cheap EM
-	private int breakdownChance	= 0;		// Chance on a d6 that a breakdown will occur when performing an HET maneuver.
-	private int bonusHets		= 0;		// Number of times this ship get's a -2 to the HET breakdown roll.
-	private int shieldCost		= 0;		// Cost to have shields active.
-	private int lifeSupportCost	= 0;		// cost to have life support active.
+	private double  movementCost	= 0;
+	private boolean nimble			= false;	// Nimble ships can do cheap EM
+	private int     breakdownChance	= 0;		// Chance on a d6 that a breakdown will occur when performing an HET maneuver.
+	private int     bonusHets		= 0;		// Number of times this ship get's a -2 to the HET breakdown roll.
+	private int     shieldCost		= 0;		// Cost to have shields active.
+	private double  lifeSupportCost	= 0;		// cost to have life support active.
 	
 	// Calculated values
-	private double hetCost		= 0;
-	private double erraticCost	= 0;
+	private double hetCost			= 0;
+	private double erraticCost		= 0;
 
 	// Real-time values
-	private int bonusHetsUsed	= 0;
+	private int bonusHetsUsed		= 0;
 
 	public PerformanceData() {
 		
@@ -31,7 +31,7 @@ public class PerformanceData {
 		breakdownChance = values.get("breakdown")   == null ? 4     : (Integer)values.get("breakdown");
 		bonusHets       = values.get("bonushets")   == null ? 0     : (Integer)values.get("bonushets");
 		shieldCost      = values.get("shieldcost")  == null ? 0     : (Integer)values.get("shieldcost");
-		lifeSupportCost = values.get("lifesupport") == null ? 0     : (Integer)values.get("lifesupport");
+		lifeSupportCost = values.get("lifesupport") == null ? 0     : (Double)values.get("lifesupport");
 		
 		// Calculate cost of HET
 		hetCost = movementCost * 5;
@@ -66,10 +66,9 @@ public class PerformanceData {
 	
 	public int getShieldCost() {
 		return this.shieldCost;
-		
 	}
 	
-	public int getLifeSupportCost() {
+	public double getLifeSupportCost() {
 		return this.lifeSupportCost;
 	}
